backend/requirements.txt
fastapi
uvicorn
pydantic-ai
python-dotenv
devtools
pydantic
logfire
duckduckgo-search
